chore: arrange detail css
