+++ src/internal/index/index.tsx
@@ -1,0 +1,3 @@
+export { default as Background } from './background';
+export { default as SkillsTable } from './skills-table';
+export { default as TopMessage } from './top-message';
