+++ src/internal/index/top-message.tsx
@@ -1,6 +1,7 @@
 import React, { useRef, useState, useEffect } from 'react';
 import { useTransition, animated } from 'react-spring';
 import styled from '@emotion/styled';
+import { StyledHppPortfolio as HppPortfolio } from './hpp-portfolio';
 
 const config = {
   from: { opacity: 0, height: 0 },
@@ -25,7 +26,7 @@ const TopMessage: React.FCX = ({ className }) => {
     ref.current.push(setTimeout(() => set([`Enjoy`]), 4000));
     ref.current.push(setTimeout(() => set([`Enjoy`, `this`]), 4750));
     ref.current.push(setTimeout(() => set([`Enjoy`, `this`, `website`]), 5250));
-    ref.current.push(setTimeout(() => set([<StyledHppPortfolio />]), 6100));
+    ref.current.push(setTimeout(() => set([<HppPortfolio />]), 6100));
     return () => ref.current.map(clearTimeout);
   }, []);
 
@@ -45,62 +46,18 @@ export const StyledTopMessage = styled(TopMessage)`
   align-items: center;
   justify-content: center;
   flex-flow: column;
-  width: 100vw;
+  width: 100%;
   height: 100vh;
 
-  padding: 5rem 10%;
   font-size: 3.5rem;
 
   @media screen and (max-width: 768px) {
-    padding: 4rem 8%;
     font-size: 3.15rem;
   }
 
   @media screen and (max-width: 480px) {
-    padding: 2.75rem 6%;
     font-size: 2.85rem;
   }
-`;
-
-const HppPortfolio: React.FCX = ({ className }) => {
-  const ref: any = useRef([]);
-  const [items, set] = useState<string[]>([]);
-  const transitions = useTransition(items, null, config);
-  const Title = 'hpp Portfolio';
-
-  useEffect(() => {
-    ref.current.map(clearTimeout);
-    ref.current = [];
-    set([]);
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 1)]), 100));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 2)]), 200));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 3)]), 300));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 4)]), 350));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 5)]), 400));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 6)]), 500));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 7)]), 600));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 8)]), 700));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 9)]), 800));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 10)]), 900));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 11)]), 1000));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 12)]), 1100));
-    ref.current.push(setTimeout(() => set([...Title.slice(0, 13)]), 1200));
-    return () => ref.current.map(clearTimeout);
-  }, []);
-
-  return (
-    <div className={className}>
-      {transitions.map(({ item, props }: any, i: number) => (
-        <animated.h3 key={i} style={props}>
-          {item}
-        </animated.h3>
-      ))}
-    </div>
-  );
-};
-
-const StyledHppPortfolio = styled(HppPortfolio)`
-  display: flex;
 `;
 
 export default StyledTopMessage;
